Change sshPublicKey in ldapPublicKey objectclass to be optional
It's quite inconvenient when the objectClass cannot be added to the user
entry until the user add a SSH key and must be removed when (s)he remove
the last SSH key. The ssh-ldap-pubkey handles this situation, but if
users don't have rights to add an objectClass, then it can't work.