find_dn_by_login: handle complex filters
RFC4515 requires filters to be wrapped with parenthesis '(' and ')'.
Over-wrapped filters are invalid! e.g. '((uid=x))'

OpenLDAP permits simple filters to omit parenthesis entirely:
e.g. 'uid=x' is automatically treated as '(uid=x)'

The OpenLDAP behavior is taken as a given in many uses, which can
lead to bad assumptions merging filters, because over-wrapped filters
ARE still rejected.

To cope with these cases, only wrap the incoming filter in
parenthesis if it does NOT already have them.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>