Fix tls_reqcert value of never to be accepted when defined
When the tls_reqcert value is set to `never` its in connect since
LDAP_OPT_X_TLS_NEVER evaluates to 0 which in turn gets evaluated to
False and the block is skipped.  I assume this is just a long standing
bug and not intended.  If the user is specifically saying they do not
care about certificate validation I would expect this to be allowed.
This commit will always set the validation settings for certificates if
the user has specified any value for tls_reqcert in the configuration.
If no value is specificed in the configuration then it continues to work
as it currently does (with default values).